Stub out nym-gateway-probe for non-unix
