+++ .github/workflows/build-nym-vpn-core-windows.yml
@@ -84,7 +84,7 @@ jobs:
       - name: Build nym-vpn-core
         working-directory: nym-vpn-core
         run: |
-          cargo build --${{ env.CARGO_TARGET }} --workspace --exclude nym-gateway-probe
+          cargo build --${{ env.CARGO_TARGET }}
 
       - name: Get rust version used for build
         id: rust-version
